create extension file to reduce size
